convertregion now supports S and H CIGAR operations
Note: this does this by just including the S and H ops in the new CIGAR string.
It is possible that this is not correct and the clipped region of the read
should be re-aligned to the converted reference position. However, that is way
out of the scope of this tool.